requirements.txt
apache-airflow>=2.5.0
cdpcli>=0.1.0 